.fix(mining): refactoring path and root dirs
